+++ tests/Nakama.Tests/AwaitedSocketTaskTest.cs
@@ -1,0 +1,54 @@
+/**
+ * Copyright 2019 The Nakama Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+using System;
+using System.Threading.Tasks;
+using Xunit;
+
+namespace Nakama.Tests
+{
+    public class AwaitedSocketTaskTest : IDisposable
+    {
+        private IClient _client;
+        private readonly ISocket _socket;
+
+        public AwaitedSocketTaskTest()
+        {
+            _client = ClientUtil.FromSettingsFile();
+            _socket = Socket.From(_client);
+        }
+
+        public void Dispose()
+        {
+            _socket.Dispose();
+            _client = null;
+        }
+
+        [Fact]
+        public async void Socket_AwaitedTasks_AreCompleted()
+        {
+            var id = Guid.NewGuid().ToString();
+            var session = await _client.AuthenticateCustomAsync(id);
+            await _socket.ConnectAsync(session);
+
+            var matchmakerTask1 = _socket.AddMatchmakerAsync("+label.foo:\"val\"", 15, 20);
+            var matchmakerTask2 = _socket.AddMatchmakerAsync("+label.bar:\"val\"", 15, 20);
+            await _socket.CloseAsync();
+
+            await Assert.ThrowsAsync<TaskCanceledException>(() => Task.WhenAll(matchmakerTask1, matchmakerTask2));
+        }
+    }
+}