+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Fixed
+- Awaited socket callback tasks are now canceled when the socket adapter is closed and cleared.
 
 ## [2.2.0] - 2019-06-06
 ### Added